// src/app/models/Cart.ts
export interface Cart {
    id: number;
    user_id: number;
    shop_id: number;
    amount: number;
    begin_amount: number;
    color: string;
    colortxt: string;
    count: number;
    domain: string;
    image: string;
    method: string;
    name: string;
    note: string;
    nv_note: string;
    kho_note: string;
    price: string;
    price_arr: string;
    pro_link: string;
    pro_properties: string;
    rate: string;
    site: string;
    size: string;
    sizetxt: string;
    status: number;
    is_deleted: number;
    created_at: string;
    updated_at: string;
}